backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
redis
aiohttp
python-dotenv
pandas
slowapi
alembic
psycopg2-binary
greenlet
